module: forward to x/mod/module (#211)
The golang.org/x/mod/module is now the canonical
location of this package.
